% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_pdf.R
\name{add_new_pdf}
\alias{add_new_pdf}
\title{Add a new PDF details into the DB.}
\usage{
add_new_pdf(table_name, data)
}
\arguments{
\item{table_name}{Name of the table to add a record to.}

\item{data}{A data.frame of the data to be added.}
}
\value{
double \code{1} if school was created. \code{0} otherwise.
}
\description{
Add a new PDF details into the DB.
}
