% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_term_end.R
\name{add_term_end}
\alias{add_term_end}
\title{admin to create current term/semester}
\usage{
add_term_end(term_end_date)
}
\arguments{
\item{term_end_date}{String. the end date for current term.
#' @return An integer indicating the number of rows affected
(typically \code{1} for a successful insert).}
}
\description{
admin to create current term/semester
}
