% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_primary_btn.R
\name{basic_primary_btn}
\alias{basic_primary_btn}
\title{Remove class 'btn-default' from a button
Add the \code{btn-primary} class}
\usage{
basic_primary_btn(btn)
}
\arguments{
\item{btn}{shiny::actionButton}
}
\description{
Remove class 'btn-default' from a button
Add the \code{btn-primary} class
}
\examples{
 btn <- shiny::actionButton(inputId = "theid", label = "The Label", class = "btn-outline-dark")
basic_primary_btn(btn)
}
