% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_payment_ticket.R
\name{create_payment_ticket}
\alias{create_payment_ticket}
\title{Add a ticket details into the DB.}
\usage{
create_payment_ticket(table_name, data)
}
\arguments{
\item{table_name}{Name of the table to add a record to.}

\item{data}{A data.frame of the data to be added.}
}
\value{
double \code{1} if school was created. \code{0} otherwise.
}
\description{
Add a ticket details into the DB.
}
