% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_school_records.R
\name{delete_school_records}
\alias{delete_school_records}
\title{Delete school records from the database}
\usage{
delete_school_records(user_id)
}
\arguments{
\item{user_id}{A character string representing the user ID}
}
\description{
Delete school records from the database
}
\examples{
# Example of deleting school records (commented out due to missing file)
# delete_school_records(user_id = "schools")
}
