% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signed_user.R
\name{get_signed_user}
\alias{get_signed_user}
\title{Get the signed in user}
\usage{
get_signed_user(user_email, user_type)
}
\arguments{
\item{user_email}{A character string representing the user email}

\item{user_type}{A character string, either \code{student} or \code{teacher}}
}
\description{
Get the signed in user
}
\examples{
email <- "useremail@gmail.com"
get_signed_user(user_email = "useremail@gmail.com", user_type = "student")
}
