% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gradebook_read}
\alias{gradebook_read}
\title{Read Class Gradebook from GoogleDrive}
\usage{
gradebook_read(file.pattern = "eGrade-BIMM143_S20", write.local.csv = FALSE)
}
\arguments{
\item{file.pattern}{Single element character vector specifying the gradebook
sheet name as stored in Drive e.g. 'eGrade-BIMM143_S20'.}

\item{write.local.csv}{Logical, if TRUE a local CSV file will be written to
the CWD.}
}
\value{
Returns a tibble/data.frame of gradebook data
}
\description{
Given an adequate file slug (i.e. file name matching pattern) this function
will go online to try and read a classes gradebook entry. It is assumed that
the start of this file (i.e. first 5 columns) are in UCSD eGrade plus email
address format that Barry and the eGrade system like to work with. This
makes final grade submission much easier
}
\examples{
 # gradebook <- gradebook_read("eGrade-BIMM143_S20")
 # head(gradebook)
}
