% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hw2gradebook}
\alias{hw2gradebook}
\title{Map HW to Gradebook Email for adding score to gradebook}
\usage{
hw2gradebook(hw, gradebook)
}
\arguments{
\item{hw}{data.frame of hw with a `Email.Address` and `Total.score` columns}

\item{gradebook}{data.frame of class gradebook entries in eGrades foramt as obtained from the functon
`gradebook_read()`.}
}
\value{
a vector of homework scores in order that matches gradebook
}
\description{
Map HW to Gradebook Email for adding score to gradebook
}
\examples{
# ToDo:
}
