% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hw_clean_quiz}
\alias{hw_clean_quiz}
\title{Tidy Quiz Questions from Google Forms}
\usage{
hw_clean_quiz(sheet, email = "Email.Address", score = "Score")
}
\arguments{
\item{sheet}{Data.frame of quiz results as obtained from `read.csv()`}

\item{email}{Column name for student Email record}

\item{score}{Column name for student final Score record}
}
\value{
A data.frame with cleaned score values
}
\description{
This function processes Barry's goggle forms quiz results. It takes a
data.frame object as input, as obtained from read.csv(), and cleans the
score column returning a data.frame with 'Email.Address', 'Score', and new
'Total.Score' columns
}
\examples{
# hw4 <- read.csv("HW4_R_quiz (Responses) - Form Responses 1.csv")
# hw_clean <- hw_clean_quiz(hw4)
# head(hw_clean)

}
