% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_production_mode.R
\name{in_development_mode}
\alias{in_development_mode}
\title{Is app running in development mode?}
\usage{
in_development_mode()
}
\value{
A length 1 logical vector. \code{TRUE} if app is in dev mode.
\code{FALSE} otherwise.
}
\description{
Another way to put it is: Is the app running locally?
}
\seealso{
\code{\link[=in_production_mode]{in_production_mode()}}
}
