% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_production_mode.R
\name{in_production_mode}
\alias{in_production_mode}
\title{Is app running in production mode?}
\usage{
in_production_mode()
}
\value{
A length 1 logical vector. \code{TRUE} if app is in production.
\code{FALSE} otherwise.
}
\description{
This function checks the \code{.Renviron} file for the variable 'in_prod'.
If \code{Sys.getenv("in_prod")} evaluates to "TRUE" or an empty string,
this function returns \code{TRUE}.
Otherwise, it returns \code{FALSE}.
}
\seealso{
\code{\link[=in_development_mode]{in_development_mode()}}
}
