% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_mandatory.R
\name{label_mandatory}
\alias{label_mandatory}
\title{add a red star after an input label}
\usage{
label_mandatory(label)
}
\arguments{
\item{label.}{A string}
}
\value{
A label with a red asterick
@examples
shiny::actionButton(inputId = "confirm_btn", label = label_mandatory("Confirm"), icon = icon("check"))
}
\description{
add a red star after an input label
}
