% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{letter.grades}
\alias{letter.grades}
\title{Vector of Letter Grades With Their Lower Bound Percent Values}
\format{
An object of class \code{numeric} of length 12.
}
\usage{
letter.grades
}
\description{
Vector of Letter Grades With Their Lower Bound Percent Values
}
\examples{
\dontrun{
 letter.grades
}
}
\keyword{datasets}
