% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{max.pts.extract}
\alias{max.pts.extract}
\title{Extract the max pts per asignment from gradebook column names}
\usage{
\method{max}{pts.extract}(gradebook)
}
\arguments{
\item{gradebook}{data.frame of class gradebook entries in eGrades foramt as obtained from the functon
`gradebook_read()`.}
}
\value{
a numeric vector of max pts with NA values for missing pts
}
\description{
This function looks for numbers within square brackets and
assumes these are the max pts for that assignment (i.e column)
}
\examples{
 gradebook.file <- "eGrade-BIMM143_F20"
 gradebook <- gradebook_read(gradebook.file)
 max.pts.extract(gradebook)

}
