% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_pdf_id.R
\name{next_pdf_id}
\alias{next_pdf_id}
\title{Get the next PDF id}
\usage{
next_pdf_id(table_name)
}
\arguments{
\item{table_name}{A character string representing the table name}
}
\description{
Get the next PDF id
}
\examples{
# next_pdf_id("pdfs")
}
