% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_request_id.R
\name{next_request_id}
\alias{next_request_id}
\title{Get the next request ID}
\usage{
next_request_id(table_name)
}
\arguments{
\item{table_name}{A character string representing the table name}
}
\description{
Get the next request ID
}
\examples{
# next_request_id("requests")
}
