% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_school_id.R
\name{next_school_id}
\alias{next_school_id}
\title{Get the next school ID}
\usage{
next_school_id(table_name)
}
\arguments{
\item{table_name}{A character string representing the table name}
}
\description{
Get the next school ID
}
\examples{
# Example of getting next school ID (commented out due to missing file)
# next_school_id("schools")
}
