% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_user_id.R
\name{next_user_id}
\alias{next_user_id}
\title{Get the next user ID}
\usage{
next_user_id(table_name, user_type)
}
\arguments{
\item{table_name}{A character string representing the table name}

\item{user_type}{A character string, either \code{student} or \code{teacher}}
}
\description{
Get the next user ID
}
\examples{
# next_user_id("users", "student")
}
