% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{percent2letter}
\alias{percent2letter}
\title{Convert percent score to letter grade}
\usage{
percent2letter(percent)
}
\arguments{
\item{percent}{a numeric vector of precent values in
the zero to 100 range.}
}
\value{
a data frame with colums for Percent and Grade
}
\description{
Convert percent score to letter grade
}
\examples{
  percent2letter( c(80, 45, 64, 99, 60) )

}
