% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_uploaded_photos.R
\name{process_uploaded_photos}
\alias{process_uploaded_photos}
\title{Process uploaded photos}
\usage{
process_uploaded_photos(photos, dest_dir, base_id)
}
\arguments{
\item{photos}{A list of file paths to the uploaded photos.}

\item{dest_dir}{The destination directory where the photos will be copied.}

\item{base_id}{The base ID to rename the photos (e.g., "REQ-001").}
}
\value{
A list of new file paths after renaming.
}
\description{
Process uploaded photos
}
