% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_bar_modified.R
\name{progress_bar_modified}
\alias{progress_bar_modified}
\title{include 100\% in the progessBar
origin function from ArgonR has excluded 0\% and 100\%
I want a full bar on last page}
\usage{
progress_bar_modified(value, text = NULL, status = "primary")
}
\description{
include 100\% in the progessBar
origin function from ArgonR has excluded 0\% and 100\%
I want a full bar on last page
}
