% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_admin_action.R
\name{record_admin_action}
\alias{record_admin_action}
\title{Capture Administrator Action and Update the Timeline Table}
\usage{
record_admin_action(user, action, description)
}
\description{
Capture Administrator Action and Update the Timeline Table
}
\examples{
\dontrun{
record_admin_action(
  user = "Admin123",
  action = "Approve",
  description = "Approved ticket #038F389E"
)
}

}
