% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_table_data.R
\name{refresh_table_data}
\alias{refresh_table_data}
\title{Refresh table data}
\usage{
refresh_table_data(table_name)
}
\arguments{
\item{table_name}{A character string representing the table name}

\item{reactive_data}{A reactive data frame}
}
\description{
Refresh table data
}
\examples{
# Example of refreshing table data (commented out due to missing file)
# refresh_table_data(reactive_data = rv$school_data, table_name = "schools")
}
