% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_new_user.R
\name{register_new_user}
\alias{register_new_user}
\title{Register a new user into the DB.
Adds a new row to the \code{schools} table in coursework database.
Status of the newly created school is "enabled" by default.}
\usage{
register_new_user(table_name, data)
}
\arguments{
\item{table_name}{Name of the table to add a record to.}

\item{data}{A data.frame of the data to be added. Colnames are
id, user_name, type, school_name, grade, phone, email, status}
}
\value{
double \code{1} if school was created. \code{0} otherwise.
}
\description{
Register a new user into the DB.
Adds a new row to the \code{schools} table in coursework database.
Status of the newly created school is "enabled" by default.
}
