% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resend_email.R
\name{resend_email}
\alias{resend_email}
\title{Resend a failed email and delete the record if successful}
\usage{
resend_email(email_id)
}
\arguments{
\item{email_id}{string. The unique identifier for the email to be resent.}
}
\value{
logical.
}
\description{
Resend a failed email and delete the record if successful
}
