% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email_notification.R
\name{send_email_notification}
\alias{send_email_notification}
\title{Function to send email and log failures to DB}
\usage{
send_email_notification(receipients, subject, body)
}
\arguments{
\item{receipients}{character. receipients email/s address/ es}

\item{subject}{string. email header}

\item{body}{html. email template used.}
}
\description{
Function to send email and log failures to DB
}
