% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_payments_status.R
\name{update_payments_status}
\alias{update_payments_status}
\title{Update the Status of a Payments in the Database}
\usage{
update_payments_status(ticket_id, new_status, balance, total, student_id)
}
\arguments{
\item{ticket_id}{A character string representing the ID of the request
that needs to be updated. Example: "TICK-007".}

\item{new_status}{A character string representing the new status to be
set for the request. Example: "PROCESSING".}
}
\description{
This function updates the status of a specific payment in the database.
}
\examples{
\dontrun{
# Example usage:
update_payments_status("TICK-007", "APPROVED")
update_payments_status("TICK-008", "CANCELLED")
}
}
