% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_school_details.R
\name{update_school_details}
\alias{update_school_details}
\title{Update school details}
\usage{
update_school_details(school_id, new_values)
}
\arguments{
\item{school_id}{A character string representing the school ID}

\item{new_values}{A named list of new values to update in the school record}
}
\description{
Update school details
}
\examples{
# update_school_details(school_id = "SCH-001", 
new_values = list(name = "New School Name", 
email = "newemail@example.com"))

}
