% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_school_status.R
\name{update_school_status}
\alias{update_school_status}
\title{Update school status}
\usage{
update_school_status(user_id, new_status)
}
\arguments{
\item{reactive_data}{A reactive data frame}

\item{table_name}{A character string representing the table name}
}
\description{
Update school status
}
\examples{
# Example of updating school status (commented out due to missing file)
# update_school_status(reactive_data = rv$school_data, table_name = "schools")
}
