% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_user_status.R
\name{update_user_status}
\alias{update_user_status}
\title{Update status in a specified table}
\usage{
update_user_status(table_name, user_id, new_status)
}
\arguments{
\item{table_name}{A character string representing the name of the table}

\item{user_id}{A character string representing the user ID}

\item{new_status}{A character string to assign status}
}
\description{
Update status in a specified table
}
\examples{
# update_status(table_name = "schools", user_id = "SCH-001", new_status = "Disabled")

}
