% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_term_end.R
\name{add_term_end}
\alias{add_term_end}
\title{Add reset payment rows for each student}
\usage{
add_term_end(term_end_date)
}
\arguments{
\item{term_end_date}{String. The end date for the current term.}
}
\value{
An integer indicating the number of rows affected.
}
\description{
Add reset payment rows for each student
}
