% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_bar_modified.R
\name{progress_bar_modified}
\alias{progress_bar_modified}
\title{include 100\% in the progessBar
origin function from ArgonR has excluded 0\% and 100\%
I want a full bar on last page}
\usage{
progress_bar_modified(value, text = NULL, status = "primary")
}
\arguments{
\item{value}{String. Numeric string to show progress}

\item{text}{String. Show label on the progress}

\item{status}{String. A color string showing the class to apply}
}
\description{
include 100\% in the progessBar
origin function from ArgonR has excluded 0\% and 100\%
I want a full bar on last page
}
\examples{
progress_bar_modified(text = page_text, value = progress_value,
status = "gradient-gray")
}
