% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_school_status.R
\name{update_school_status}
\alias{update_school_status}
\title{Update school status}
\usage{
update_school_status(user_id, new_status)
}
\arguments{
\item{user_id}{A character string representing the user ID}

\item{new_status}{A character string to asign status}
}
\description{
Update school status
}
\examples{
# update_school_status(user_id = "SCH-001", new_status = "Disabled")

}
